requirements.txt
# UFDR Analysis Tool - Dependencies
# Offline-first RAG architecture

# === Frontend ===
streamlit>=1.28.0

# === RAG Engine (all local, no API key needed) ===
chromadb>=0.4.0                   # Local vector store + sentence-transformers
rank_bm25>=0.2.2                  # BM25 keyword search (pure Python)

# === LLM API (only cloud component, needs API key) ===
google-generativeai>=0.3.0        # Gemini API for reasoning/reports
openai>=1.0.0                       # OpenAI/OpenRouter client

# === Database ===
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# === Data Processing ===
pandas>=2.0.0
numpy>=1.24.0

# === Core Utilities ===
python-dateutil>=2.8.0
phonenumbers>=8.13.0
pydantic>=2.0.0

# === XML Processing ===
lxml>=4.9.0
xmltodict>=0.13.0

# === Testing ===
pytest>=7.0.0
pytest-cov>=4.0.0
